[VPO] Fix type mismatch for cost calculation and returned value.
This fixes failure on Transforms/Intel_VPO/Vecopt/hir_cost_overflow.ll
test that was introduced during merge from xmain to vpo-xmain.

There are some more suspicious cost-related variables that are
"unsigned int" instead of "uint64_t" (in which most calculation are
done and which is returned) but it's less worse than signed/unsigned
type mismatch that this commit fixes and is left for future cleanup.

Old/split llvm commit ID: 8c4587438cc50e7c62bb10f8c9069ad84536ebd9

+++ llvm/include/llvm/Analysis/Intel_VPO/Vecopt/VPOScenarioEvaluation.h
@@ -468,8 +468,8 @@ public:
   // want to encode the results of the CostModel, namely, which ALoops in the
   // region to vectorize and using which VFs (directly/explicitely in the
   // AVR?...)
-  int getCost(AVRLoop *ALoop, unsigned int VF, VPOVLSInfoBase *VLSInfo,
-              HIRSafeReductionAnalysis *SRA);
+  uint64_t getCost(AVRLoop *ALoop, unsigned int VF, VPOVLSInfoBase *VLSInfo,
+                   HIRSafeReductionAnalysis *SRA);
 
   virtual VPOCostGathererBase *getCostGatherer(unsigned int VF, AVRLoop *ALoop, 
                                                VPOVLSInfoBase *VLSInfo,
